requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
yt-dlp==2023.11.16
openai-whisper==20231117
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
supabase==2.12.0
pinecone-client==3.0.0 